# style: use for..of