--- a/lib/extend/filter.js
+++ b/lib/extend/filter.js
@@ -71,8 +71,8 @@ Filter.prototype.execSync = function(type, data, options = {}) {
 
   args.unshift(data);
 
-  for (let i = 0, len = filters.length; i < len; i++) {
-    const result = Reflect.apply(filters[i], ctx, args);
+  for (const filter of filters) {
+    const result = Reflect.apply(filter, ctx, args);
     args[0] = result == null ? args[0] : result;
   }