--- a/lib/hexo/index.js
+++ b/lib/hexo/index.js
@@ -44,15 +44,13 @@ const createLoadThemeRoute = function(generatorResult, locals, ctx) {
   const { path, cache: useCache } = locals;
 
   const layout = unique(castArray(generatorResult.layout));
-  const layoutLength = layout.length;
 
   // allways use cache in fragment_cache
   locals.cache = true;
   return () => {
     if (useCache && routeCache.has(generatorResult)) return routeCache.get(generatorResult);
 
-    for (let i = 0; i < layoutLength; i++) {
-      const name = layout[i];
+    for (const name of layout) {
       const view = theme.getView(name);
 
       if (view) {
@@ -411,8 +409,8 @@ class Hexo {
       // Remove old routes
       const removed = routeList.filter(item => !newRouteList.includes(item));
 
-      for (let i = 0, len = removed.length; i < len; i++) {
-        route.remove(removed[i]);
+      for (const i of removed) {
+        route.remove(i);
       }
     });
   }