--- a/lib/hexo/locals.js
+++ b/lib/hexo/locals.js
@@ -52,8 +52,7 @@ Locals.prototype.toObject = function() {
   const result = {};
   const keys = Object.keys(this.getters);
 
-  for (let i = 0, len = keys.length; i < len; i++) {
-    const key = keys[i];
+  for (const key of keys) {
     const item = this.get(key);
 
     if (item != null) result[key] = item;