--- a/lib/hexo/multi_config_path.js
+++ b/lib/hexo/multi_config_path.js
@@ -30,22 +30,20 @@ module.exports = ctx => function multiConfigPath(base, configPaths, outputDir) {
     return configPath;
   }
 
-  const numPaths = paths.length;
-
   // combine files
   const combinedConfig = {};
   let count = 0;
-  for (let i = 0; i < numPaths; i++) {
-    const configPath = isAbsolute(paths[i]) ? paths[i] : join(base, paths[i]);
+  for (const path of paths) {
+    const configPath = isAbsolute(path) ? path : join(base, path);
 
     if (!fs.existsSync(configPath)) {
-      log.w(`Config file ${paths[i]} not found.`);
+      log.w(`Config file ${path} not found.`);
       continue;
     }
 
     // files read synchronously to ensure proper overwrite order
     const file = fs.readFileSync(configPath);
-    const ext = extname(paths[i]).toLowerCase();
+    const ext = extname(path).toLowerCase();
 
     if (ext === '.yml') {
       merge(combinedConfig, yml.load(file));
@@ -54,7 +52,7 @@ module.exports = ctx => function multiConfigPath(base, configPaths, outputDir) {
       merge(combinedConfig, yml.safeLoad(file, {json: true}));
       count++;
     } else {
-      log.w(`Config file ${paths[i]} not supported type.`);
+      log.w(`Config file ${path} not supported type.`);
     }
   }