--- a/lib/hexo/register_models.js
+++ b/lib/hexo/register_models.js
@@ -7,8 +7,7 @@ module.exports = ctx => {
 
   const keys = Object.keys(models);
 
-  for (let i = 0, len = keys.length; i < len; i++) {
-    const key = keys[i];
+  for (const key of keys) {
     db.model(key, models[key](ctx));
   }
 };