--- a/lib/plugins/console/config.js
+++ b/lib/plugins/console/config.js
@@ -53,8 +53,7 @@ function setProperty(obj, key, value) {
   let cursor = obj;
   const lastKey = split.pop();
 
-  for (let i = 0, len = split.length; i < len; i++) {
-    const name = split[i];
+  for (const name of split) {
     cursor[name] = cursor[name] || {};
     cursor = cursor[name];
   }