--- a/lib/plugins/console/list/route.js
+++ b/lib/plugins/console/list/route.js
@@ -19,12 +19,11 @@ function buildTree(routes) {
   const obj = {};
   let cursor;
 
-  for (let i = 0, len = routes.length; i < len; i++) {
-    const item = routes[i].split('/');
+  for (const route of routes) {
+    const item = route.split('/');
     cursor = obj;
 
-    for (let j = 0, lenj = item.length; j < lenj; j++) {
-      const seg = item[j];
+    for (const seg of item) {
       cursor[seg] = cursor[seg] || {};
       cursor = cursor[seg];
     }
@@ -37,8 +36,7 @@ function buildNodes(tree) {
   const keys = Object.keys(tree);
   const nodes = [];
 
-  for (let i = 0, len = keys.length; i < len; i++) {
-    const key = keys[i];
+  for (const key of keys) {
     const item = tree[key];
 
     if (Object.keys(item).length) {