--- a/lib/plugins/console/new.js
+++ b/lib/plugins/console/new.js
@@ -35,8 +35,7 @@ function newConsole(args) {
 
   const keys = Object.keys(args);
 
-  for (let i = 0, len = keys.length; i < len; i++) {
-    const key = keys[i];
+  for (const key of keys) {
     if (!reservedKeys[key]) data[key] = args[key];
   }