--- a/lib/plugins/filter/new_post_path.js
+++ b/lib/plugins/filter/new_post_path.js
@@ -67,8 +67,7 @@ function newPostPathFilter(data = {}, replace) {
           title: slug
         };
 
-        for (let i = 0, len = keys.length; i < len; i++) {
-          const key = keys[i];
+        for (const key of keys) {
           if (!reservedKeys[key]) filenameData[key] = data[key];
         }