--- a/lib/plugins/processor/post.js
+++ b/lib/plugins/processor/post.js
@@ -59,8 +59,7 @@ module.exports = ctx => {
         data.published = false;
       }
 
-      for (let i = 0, len = keys.length; i < len; i++) {
-        const key = keys[i];
+      for (const key of keys) {
         if (!preservedKeys[key]) data[key] = info[key];
       }