@segayuu 2019-11-17
https://jsbench.me/o2k334c1v1/1
As a result of running to the end on one site, it was about 40% slower than Classic for loop.
Personally, Classic for loop is not difficult to read. Based on this result, I do not agree with this PR.
However, since the speed does not change as much as IO processing, it may be ignored. Honest preference :)