@SukkaW 2019-11-17
@segayuu I don't know why `for..of` is slower in browser.

According to a benchmark run on runkit, it seems `for..of` is much more faster: https://runkit.com/sukkaw/5dd164a025808400140b9276