@curbengh 2019-11-18
I do agree with the classic style is easier to read as it's similar to other languages, whereas `for...of` could be confused with `for...in`.

But the node benchmark looks very convincing though, can you setup another benchmark on Node 10/12?