@curbengh 2019-11-22
RunKit somehow doesn't run benchmark anymore. Anyhow, I ran it using my machine:

```
Node 8.16
Run 1
for x 55.37 ops/sec ±2.81% (68 runs sampled)
for len x 49.25 ops/sec ±3.26% (63 runs sampled)
for of x 32.80 ops/sec ±2.86% (57 runs sampled)
Run 2
for x 55.17 ops/sec ±3.01% (60 runs sampled)
for len x 49.64 ops/sec ±3.03% (63 runs sampled)
for of x 31.18 ops/sec ±3.96% (55 runs sampled)
Run 3
for x 56.00 ops/sec ±2.88% (69 runs sampled)
for len x 49.54 ops/sec ±3.17% (63 runs sampled)
for of x 32.10 ops/sec ±3.28% (55 runs sampled)
```

```
Node 10.16.1
Run 1
for x 70.97 ops/sec ±2.98% (70 runs sampled)
for len x 60.21 ops/sec ±3.37% (61 runs sampled)
for of x 7.37 ops/sec ±1.16% (23 runs sampled)
Run 2
for x 71.19 ops/sec ±3.13% (70 runs sampled)
for len x 61.12 ops/sec ±3.11% (62 runs sampled)
for of x 7.34 ops/sec ±3.29% (23 runs sampled)
Run 3
for x 70.79 ops/sec ±3.13% (70 runs sampled)
for len x 60.70 ops/sec ±3.68% (62 runs sampled)
for of x 7.16 ops/sec ±3.81% (22 runs sampled)
```