@SukkaW 2019-11-22
@curbengh It seems unnormal. At least `for len` should be a little bit faster than `for` because array.length won't be calculated again and again.

Here is the benchmark on repl.it, and `for..of` is the slowest.

https://repl.it/repls/FavorableSqueakyProblems

```
for x 18.58 ops/sec ±5.42% (33 runs sampled)
for len x 18.76 ops/sec ±5.06% (34 runs sampled)
for of x 2.15 ops/sec ±3.69% (10 runs sampled)
```