@curbengh 2019-11-22
I also tried separating them into individual files, I ran again in my machine,

```
Node 10
for of x 71.67 ops/sec ±3.22% (71 runs sampled)
for of x 71.97 ops/sec ±3.24% (71 runs sampled)
for of x 66.64 ops/sec ±6.05% (60 runs sampled)
```

```
for len x 59.19 ops/sec ±6.09% (59 runs sampled)
for len x 57.70 ops/sec ±4.44% (59 runs sampled)
for len x 60.74 ops/sec ±3.19% (61 runs sampled)
```

```
for x 7.01 ops/sec ±6.44% (21 runs sampled)
for x 6.82 ops/sec ±7.52% (21 runs sampled)
for x 6.04 ops/sec ±10.92% (20 runs sampled)
```

Looks like there is a regression in Node 10.