@SukkaW 2019-12-01
I have rebased this branch up to latest master. Here is the benchmark result:

| | Node.js 8 | Node.js 10 | Node.js 12 | Node.js 13 |
|:---:|:---:|:---:|:---:|:---:|
| Current master branch | 29.8s | 25.6s | 24.5s | 26.6s |
| This PR | 31.5s | 26.5s | 27.3s | 27.1s |

This PR is indeed a little bit alower.